Fix imports

+++ src/models/core/player_info/PlayerInfo.py
@@ -1,4 +1,5 @@
 from pydantic import BaseModel
 
+
 class PlayerInfo(BaseModel):
-  bankroll: int
+  bankroll: float
